% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{infoDataSet}
\alias{infoDataSet}
\title{Retrieve Dataset Information}
\usage{
infoDataSet(dataset)
}
\arguments{
\item{dataset}{A dataframe or list containing dataset attributes.}
}
\value{
A list with the extracted dataset attributes, including:
\item{id}{Dataset ID}
\item{name}{Dataset name}
\item{instances}{Number of instances}
\item{inputs}{Number of input attributes}
\item{labels}{Number of labels}
\item{LabelsSets}{Label sets information}
\item{single}{Single-label classification indicator}
\item{maxfreq}{Maximum frequency of labels}
\item{card}{Cardinality of the dataset}
\item{dens}{Density of the dataset}
\item{mean}{Mean value of dataset attributes}
\item{scumble}{Scumble measure}
\item{tcs}{TCS value (Label co-occurrence complexity)}
\item{attStart}{Starting index of attributes}
\item{attEnd}{Ending index of attributes}
\item{labStart}{Starting index of labels}
\item{labEnd}{Ending index of labels}
\item{distinct}{Number of distinct values}
\item{xn, yn, gridn, xt, yt, gridt}{Grid-related attributes}
}
\description{
This function extracts information from a dataset object, typically from a file
like "datasets-hpmlk.csv". It returns various attributes related to the dataset.
}
\examples{
dataset <- data.frame(ID = 1, Name = "SampleData", Instances = 100, Inputs = 10, Labels = 5)
info <- infoDataSet(dataset)
print(info$name)  # Output: "SampleData"
}
