% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testMaF1.R
\name{maf1.test.build.data.frame}
\alias{maf1.test.build.data.frame}
\title{Build an Empty Data Frame for MAF1 Test}
\usage{
maf1.test.build.data.frame()
}
\value{
A data frame with predefined evaluation measures as the first column.
}
\description{
This function initializes an empty data frame with predefined evaluation measures as row names.
}
\details{
\itemize{
\item Creates a 22-row by 11-column data frame initialized with NA values.
\item The first column is populated with evaluation measure names.
}
}
