% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mergeAndSaveResults}
\alias{mergeAndSaveResults}
\title{Merge and Save Performance Results}
\usage{
mergeAndSaveResults(
  roc_auc_1,
  roc_auc_macro_1,
  roc_auc_micro_1,
  auprc_proba_1,
  resultados_proba_1
)
}
\arguments{
\item{roc_auc_1}{A list containing ROC AUC values for different KNN models.}

\item{roc_auc_macro_1}{A list containing macro ROC AUC values for different KNN models.}

\item{roc_auc_micro_1}{A list containing micro ROC AUC values for different KNN models.}

\item{auprc_proba_1}{A list containing AUPRC values (both micro and macro) for different KNN models.}

\item{resultados_proba_1}{A list containing probability-based results for different KNN models.}
}
\value{
This function does not return a value; it writes merged results to CSV files.
}
\description{
This function merges multiple performance metrics (ROC AUC, AUPRC, etc.) into a single dataframe
and saves it as a CSV file.
}
\examples{
\dontrun{
mergeAndSaveResults(roc_auc_1, roc_auc_macro_1, roc_auc_micro_1, auprc_proba_1, resultados_proba_1)
}
}
