% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{roc.curva}
\alias{roc.curva}
\title{Compute and Plot ROC Curve for Multi-Label Classification}
\usage{
\method{roc}{curva}(f, y_pred, test, Folder, nome)
}
\arguments{
\item{f}{An integer indicating the fold number (useful for cross-validation).}

\item{y_pred}{A vector or dataframe containing the predicted probabilities.}

\item{test}{A dataframe containing the true labels (ground truth).}

\item{Folder}{A string specifying the directory where the ROC plot and metrics should be saved.}

\item{nome}{A string used to name the saved files.}
}
\value{
A list containing:
\item{roc_auc}{A dataframe with the ROC AUC value.}
\item{roc_auc_macro}{A dataframe with the macro-averaged ROC AUC value.}
\item{roc_auc_micro}{A dataframe with the micro-averaged ROC AUC value.}
}
\description{
This function calculates the ROC (Receiver Operating Characteristic) curve,
plots it, and computes the AUC (Area Under the Curve) for multi-label classification.
}
\examples{
\dontrun{
result <- roc.curva(1, predicted_probs, test_labels, "results", "model_roc")
}
}
