% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSilho.R
\name{silho.build.data.frame}
\alias{silho.build.data.frame}
\title{Build an Empty Data Frame for Silhouette Evaluation}
\usage{
silho.build.data.frame()
}
\value{
A data frame with 22 rows and 11 columns, where the first column is filled with measure names.
}
\description{
This function creates an empty data frame with predefined evaluation measures.
}
