% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSilho.R
\name{silho.gather.evaluated}
\alias{silho.gather.evaluated}
\title{Gather Evaluated Silhouette Test Results}
\usage{
silho.gather.evaluated(parameters)
}
\arguments{
\item{parameters}{A list containing necessary settings and paths, including:
\itemize{
\item \code{Number.Folds}: The number of folds to process.
\item \code{Folders$folderPartitions}: Path to partition folders.
\item \code{Folders$folderTestSilho}: Path to test silhouette folders.
\item \code{Choosed$all.methods.choosed}: Data frame of selected methods.
}}
}
\value{
None. The function writes aggregated evaluation results to CSV files.
}
\description{
This function collects and organizes evaluation results from multiple folds and KNN runs.
}
