% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{splitAuprc}
\alias{splitAuprc}
\title{Split AUPRC DataFrame by KNN Models}
\usage{
splitAuprc(df, nome_base)
}
\arguments{
\item{df}{A dataframe containing AUPRC metrics with columns: "fold", "Knn", "Micro.AUPRC", and "Macro.AUPRC".}

\item{nome_base}{A base name (string) for identification purposes (not used in function logic).}
}
\value{
A named list of dataframes, where each element contains AUPRC data for a specific KNN model and type (Micro/Macro).
}
\description{
This function splits an AUPRC dataframe into a list of dataframes,
where each element corresponds to a specific KNN model and AUPRC type (Micro/Macro).
}
\examples{
\dontrun{
split_auprc_results <- splitAuprc(df, "my_dataset")
}
}
