% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSilho.R
\name{testSilhouette}
\alias{testSilhouette}
\title{Run the Silhouette Test Workflow}
\usage{
testSilhouette(parameters)
}
\arguments{
\item{parameters}{A list containing necessary settings and paths, including:
\itemize{
\item \code{Folders$folderTestSilho}: Path to store test silhouette results.
\item \code{Dataset.Name}: Name of the dataset being processed.
}}
}
\value{
None. The function writes runtime statistics to a CSV file.
}
\description{
This function executes the full Silhouette test workflow, including building, testing, evaluating, and organizing partitions.
}
